CPP: Make libphonenumber depend on its own base/ when it's used in Chromium.
This is done by adding a USE_BOOST compile time flag that is enabled by default
when the library is compiled with CMake (mainly for external users).

In Chromium, the library will be built without the USE_BOOST compile time flag.
Instead the NO_THREAD_SAFETY flag will have to be provided so that the minimal
non-thread-safe subset of base will be used.
This is possible since libphonenumber is only used from the UI thread in
Chromium.

BUG=http://crbug.com/236272
R=jia.shao.peng@gmail.com

Review URL: https://codereview.appspot.com/8859052

git-svn-id: http://libphonenumber.googlecode.com/svn@568 ee073f10-1060-11df-b6a4-87a95322a99c